Add validations for IWA headers
This commit adds the support for checking that all the required IWA headers are
in place for IWAs. It also adds an attribute `integrityBlockSign.isIwa` to be
able to disable those checks if in the future non-iwas will be built with the
tool. The value will for now default to true.